Move indicators graphql folders/file to correct place
