--- server/graphql/programs/src/mutations/program_indicator/mod.rs
@@ -1,1 +1,0 @@
-pub mod update_value;
