+++ server/graphql/requisition/src/mutations/mod.rs
@@ -1,3 +1,4 @@
 pub mod errors;
 pub mod request_requisition;
 pub mod response_requisition;
+pub mod update_indicator_value;
